# src/main/resources/static/api-docs.yaml
openapi: 3.0.3
info:
  title: Cinema Ticketing API
  version: 2.0.0
  description: API для системы продажи билетов в кинотеатр.
servers:
  - url: http://91.142.94.183:8080
  - url: http://localhost:8080
paths:
...
  /users/me:
    get:
      summary: Получить текущего пользователя
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Не авторизован

    put:
      summary: Обновить данные пользователя
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Данные обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации
  /users/{id}:
    get:
      summary: Получить пользователя по ID (admin)
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Пользователь не найден
...
  /sessions:
    get:
      summary: Получить список сеансов
      tags: [Sessions]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: filmId
          in: query
          description: Фильтр по ID фильма
          schema:
            type: string
        - name: date
          in: query
          description: Фильтр по дате (YYYY-MM-DD)
          schema:
            type: string
      responses:
        '200':
          description: Список сеансов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
    post:
      summary: Создать сеанс (admin) (periodicConfig = null для одиночного сеанса)
      tags: [Sessions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreate'
      responses:
        '201':
          description: Сеанс создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
  /sessions/{id}:
    get:
      summary: Получить сеанс по ID
      tags: [Sessions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Сеанс
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          description: Сеанс не найден

    put:
      summary: Обновить сеанс (admin)
      tags: [Sessions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionUpdate'
      responses:
        '200':
          description: Сеанс обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          description: Сеанс не найден

    delete:
      summary: Удалить сеанс (admin)
      tags: [Sessions]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Сеанс удален
        '404':
          description: Сеанс не найден
...
  /seat-categories:
    get:
      summary: Получить список категорий мест
      tags: [SeatCategories]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Список категорий мест
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatCategoryListResponse'
    post:
      summary: Создать категорию мест (admin)
      tags: [SeatCategories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatCategoryCreate'
      responses:
        '201':
          description: Категория создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatCategory'
  /seat-categories/{id}:
    get:
      summary: Получить категорию мест по ID
      tags: [SeatCategories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Категория мест
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatCategory'
        '404':
          description: Категория не найдена

    put:
      summary: Обновить категорию мест (admin)
      tags: [SeatCategories]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatCategoryUpdate'
      responses:
        '200':
          description: Категория обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatCategory'
        '404':
          description: Категория не найдена

    delete:
      summary: Удалить категорию мест (admin)
      tags: [SeatCategories]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Категория удалена
        '404':
          description: Категория не найдена
  /sessions/{sessionId}/tickets:
    get:
      summary: Получить билеты для сеанса
      tags: [Tickets]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          description: Фильтр по статусу билета
          schema:
            type: string
            enum: [AVAILABLE, RESERVED, SOLD, CANCELLED]
      responses:
        '200':
          description: Список билетов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '404':
          description: Сеанс не найден
  /tickets/{id}/reserve:
    post:
      summary: Забронировать билет
      tags: [Tickets]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Билет забронирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          description: Билет недоступен для бронирования
        '404':
          description: Билет не найден
  /tickets/{id}/cancel-reservation:
    post:
      summary: Отменить бронирование билета
      tags: [Tickets]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Бронирование отменено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          description: Невозможно отменить бронирование
        '404':
          description: Билет не найден
...
  /films/{filmId}/reviews:
    get:
      summary: Получить отзывы на фильм
      tags: [Reviews]
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Список отзывов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '404':
          description: Фильм не найден

    post:
      summary: Оставить отзыв на фильм
      tags: [Reviews]
      security:
        - bearerAuth: []
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: Отзыв создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Ошибка валидации или нет покупки на этот фильм
        '404':
          description: Фильм не найден
  /reviews/{id}:
    get:
      summary: Получить отзыв по ID
      tags: [Reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Отзыв
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Отзыв не найден

    put:
      summary: Обновить свой отзыв
      tags: [Reviews]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdate'
      responses:
        '200':
          description: Отзыв обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '403':
          description: Нет прав для редактирования
        '404':
          description: Отзыв не найден

    delete:
      summary: Удалить отзыв
      tags: [Reviews]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Отзыв удален
        '403':
          description: Нет прав для удаления
        '404':
          description: Отзыв не найден
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        age: { type: integer }
        role:
          type: string
          enum: [USER, ADMIN]
        gender:
          type: string
          enum: [MALE, FEMALE]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UserUpdate:
      type: object
      properties:
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        age: { type: integer }
        gender:
          type: string
          enum: [MALE, FEMALE]
...
    SessionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        pages: { type: integer }
...
    SessionCreate:
      type: object
      properties:
        filmId: { type: string }
        hallId: { type: string }
        startAt: { type: string, format: date-time }
        periodicConfig:
          $ref: '#/components/schemas/PeriodicSessionConfig'
      required: [filmId, hallId, startAt]
    PeriodicSessionConfig:
      type: object
      properties:
        period:
          type: string
          enum: [EVERY_DAY, EVERY_WEEK]
        periodGenerationEndsAt: { type: string, format: date-time }
      required: [period, periodGenerationEndsAt]
...
    Session:
      type: object
      properties:
        id: { type: string }
        filmId: { type: string }
        hallId: { type: string }
        startAt:
          type: string
          format: date-time
        timeslot:
          type: object
          properties:
            start: { type: string, format: date-time }
            end: { type: string, format: date-time }
    Ticket:
      type: object
      properties:
        id: { type: string }
        sessionId: { type: string }
        seatId: { type: string }
        categoryId: { type: string }
        priceCents: { type: integer }
        status:
          type: string
          enum: [AVAILABLE, RESERVED, SOLD, CANCELLED]
        reservedUntil: { type: string, format: date-time }
...
    Review:
      type: object
      properties:
        id: { type: string }
        filmId: { type: string }
        clientId: { type: string }
        rating: { type: integer, minimum: 0, maximum: 5 }
        text: { type: string }
        createdAt: { type: string, format: date-time }
...
    SeatCategory:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        priceCents: { type: integer }
...
    SessionUpdate:
      type: object
      properties:
        filmId: { type: string }
        hallId: { type: string }
        startAt: { type: string, format: date-time }
...
    SeatCategoryListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SeatCategory'
        pagination:
          $ref: '#/components/schemas/Pagination'
    SeatCategoryCreate:
      type: object
      properties:
        name: { type: string }
        priceCents: { type: integer }
      required: [name, priceCents]
    SeatCategoryUpdate:
      type: object
      properties:
        name: { type: string }
        priceCents: { type: integer }
...
    ReviewListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ReviewCreate:
      type: object
      properties:
        rating: { type: integer, minimum: 0, maximum: 5 }
        text: { type: string }
      required: [rating, text]
    ReviewUpdate:
      type: object
      properties:
        rating: { type: integer, minimum: 0, maximum: 5 }
        text: { type: string }
...
  parameters:
    page:
      name: page
      in: query
      description: Номер страницы
      schema:
        type: integer
        default: 0
        minimum: 0
    limit:
      name: size
      in: query
      description: Количество элементов на странице
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT